refactor: Application 네이밍 수정
