+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'couponpop-member'
+rootProject.name = 'member-service'
